'prueba'

+++ backend/ms-productos/src/main/java/isi/dan/ms_productos/controller/ProductoController.java
@@ -1,21 +1,28 @@
 package isi.dan.ms_productos.controller;
 
+import java.util.List;
+import java.util.Optional;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.client.RestTemplate;
-
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import isi.dan.ms_productos.aop.LogExecutionTime;
 import isi.dan.ms_productos.exception.ProductoNotFoundException;
+import isi.dan.ms_productos.modelo.Categoria;
 import isi.dan.ms_productos.modelo.Producto;
 import isi.dan.ms_productos.servicio.EchoClientFeign;
 import isi.dan.ms_productos.servicio.ProductoService;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/productos")
@@ -36,23 +43,23 @@ public class ProductoController {
         return ResponseEntity.ok(savedProducto);
     }
 
-    @GetMapping("/test")
-    @LogExecutionTime
-    public String getEcho() {
-        String resultado = echoSvc.echo();
-        log.info("Log en test 1!!!! {}",resultado);
-        return resultado;
-    }
+    // @GetMapping("/test")
+    // @LogExecutionTime
+    // public String getEcho() {
+    //     String resultado = echoSvc.echo();
+    //     log.info("Log en test 1!!!! {}",resultado);
+    //     return resultado;
+    // }
 
-    @GetMapping("/test2")
-    @LogExecutionTime
-    public String getEcho2() {
-        RestTemplate restTemplate = new RestTemplate();
-        String gatewayURL = "http://ms-gateway-svc:8080";
-        String resultado = restTemplate.getForObject(gatewayURL+"/clientes/api/clientes/echo", String.class);
-        log.info("Log en test 2 {}",resultado);
-        return resultado;
-    }
+    // @GetMapping("/test2")
+    // @LogExecutionTime
+    // public String getEcho2() {
+    //     RestTemplate restTemplate = new RestTemplate();
+    //     String gatewayURL = "http://ms-gateway-svc:8080";
+    //     String resultado = restTemplate.getForObject(gatewayURL+"/clientes/api/clientes/echo", String.class);
+    //     log.info("Log en test 2 {}",resultado);
+    //     return resultado;
+    // }
 
     @GetMapping
     @LogExecutionTime
@@ -62,13 +69,49 @@ public class ProductoController {
 
     @GetMapping("/{id}")
     @LogExecutionTime
-    public ResponseEntity<Producto> getProductoById(@PathVariable Long id) throws ProductoNotFoundException {
-        return  ResponseEntity.ok(productoService.getProductoById(id));
+    public ResponseEntity<Producto> getProductoById(@PathVariable Long id) {
+        Optional<Producto> producto = productoService.getProductoById(id);
+        return producto.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+    }
+
+    // Get by Categoria
+    @GetMapping("/categoria/{categoria}")
+    @LogExecutionTime
+    public List<Producto> getProductosByCategoria(@PathVariable String categoria) throws ProductoNotFoundException {
+        try {
+            Categoria cat = Categoria.valueOf(categoria.toUpperCase());
+            return productoService.getProductosByCategoria(cat);
+        } catch (IllegalArgumentException e) {
+            throw new ProductoNotFoundException(categoria);
+        }
     }
 
     @DeleteMapping("/{id}")
     @LogExecutionTime
     public ResponseEntity<Void> deleteProducto(@PathVariable Long id) {
+        productoService.deleteProducto(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    // Delete by Categoria
+    @DeleteMapping("/categoria/{categoria}")
+    @LogExecutionTime
+    public ResponseEntity<Void> deleteByCategoria(@PathVariable String categoria) throws ProductoNotFoundException {
+        try {
+            Categoria cat = Categoria.valueOf(categoria.toUpperCase());
+            productoService.deleteByCategoria(cat);
+            return ResponseEntity.noContent().build();
+        } catch(IllegalArgumentException e ){
+            throw new ProductoNotFoundException(categoria);
+        }
+    }
+
+    @PutMapping("/{id}")
+    @LogExecutionTime
+    public ResponseEntity<Producto> updateProducto(@PathVariable Long id, @RequestBody Producto producto) throws ProductoNotFoundException{
+        if(!productoService.getProductoById(id).isPresent()){
+            return ResponseEntity.notFound().build();
+        }
         productoService.deleteProducto(id);
         return ResponseEntity.noContent().build();
     }
